chore: move logger to infra directory and update import path
